improv: layout button styling
