+++ src/pages/color_schemes/preview.rs
@@ -1,9 +1,7 @@
 use crate::fl;
 use cosmic::{
-    cosmic_theme::Theme,
-    iced::{Alignment, Border, Color, Length},
-    iced_core::Shadow,
-    widget::{self, button, container, tooltip},
+    iced::{Alignment, Length},
+    widget::{self, tooltip},
     Apply, Element,
 };
 
@@ -32,7 +30,7 @@ pub fn installed<'a>(
                     .padding(spacing.space_xxs)
                     .width(Length::Fixed(100.0))
                     .height(Length::Fill)
-                    .class(card(theme.clone()))
+                    .class(crate::app::style::card(theme.clone()))
                     .into(),
                 widget::horizontal_space().into(),
                 widget::tooltip::tooltip(
@@ -64,7 +62,7 @@ pub fn installed<'a>(
         .width(Length::Fixed(200.0))
         .height(Length::Fixed(160.0))
         .apply(widget::container)
-        .class(background(theme.clone())),
+        .class(crate::app::style::background(theme.clone())),
     )
     .selected(selected.name == color_scheme.name)
     .class(cosmic::style::Button::Image)
@@ -94,7 +92,7 @@ pub fn available<'a>(color_scheme: &ColorScheme) -> Element<'a, crate::app::Mess
                     .padding(spacing.space_xxs)
                     .width(Length::Fixed(100.0))
                     .height(Length::Fill)
-                    .class(card(theme.clone()))
+                    .class(crate::app::style::card(theme.clone()))
                     .into(),
                 widget::horizontal_space().into(),
                 widget::tooltip::tooltip(
@@ -129,81 +127,11 @@ pub fn available<'a>(color_scheme: &ColorScheme) -> Element<'a, crate::app::Mess
         ])
         .height(Length::Fixed(160.0))
         .apply(widget::container)
-        .class(background(theme.clone())),
+        .class(crate::app::style::background(theme.clone())),
     )
     .class(cosmic::style::Button::Image)
     .on_press(crate::app::Message::ColorSchemes(Box::new(
         super::Message::SetColorScheme(color_scheme.clone()),
     )))
     .into()
-}
-
-pub fn background<'a>(theme: Theme) -> cosmic::theme::Container<'a> {
-    let corner_radii = cosmic::theme::active().cosmic().corner_radii;
-    cosmic::theme::Container::custom(move |_| container::Style {
-        icon_color: Some(Color::from(theme.background.on)),
-        text_color: Some(Color::from(theme.background.on)),
-        background: Some(cosmic::iced::Background::Color(
-            theme.background.base.into(),
-        )),
-        border: Border {
-            radius: corner_radii.radius_xs.into(),
-            ..Default::default()
-        },
-        shadow: Shadow::default(),
-    })
-}
-
-pub fn card<'a>(theme: Theme) -> cosmic::theme::Container<'a> {
-    let theme = theme.clone();
-    let corner_radii = cosmic::theme::active().cosmic().corner_radii;
-
-    cosmic::theme::Container::custom(move |_| container::Style {
-        icon_color: Some(Color::from(theme.primary.component.on)),
-        text_color: Some(Color::from(theme.primary.component.on)),
-        background: Some(cosmic::iced::Background::Color(
-            theme.primary.component.base.into(),
-        )),
-        border: Border {
-            radius: corner_radii.radius_s.into(),
-            ..Default::default()
-        },
-        shadow: Shadow::default(),
-    })
-}
-
-#[allow(dead_code)]
-pub fn standard_button(theme: Theme) -> cosmic::theme::Button {
-    let active_theme = theme.clone();
-    let disabled_theme = theme.clone();
-    let hovered_theme = theme.clone();
-    let pressed_theme = theme.clone();
-    let _corner_radii = cosmic::theme::active().cosmic().corner_radii;
-
-    cosmic::theme::Button::Custom {
-        active: Box::new(move |_active, _cosmic| button::Style {
-            background: Some(cosmic::iced_core::Background::Color(
-                active_theme.on_accent_color().into(),
-            )),
-            ..Default::default()
-        }),
-        disabled: Box::new(move |_cosmic| button::Style {
-            background: Some(cosmic::iced_core::Background::Color(
-                disabled_theme.on_accent_color().into(),
-            )),
-            ..Default::default()
-        }),
-        hovered: Box::new(move |_hovered, _cosmic| button::Style {
-            background: Some(cosmic::iced_core::Background::Color(
-                hovered_theme.on_accent_color().into(),
-            )),
-            ..Default::default()
-        }),
-        pressed: Box::new(move |_pressed, _cosmic| button::Style {
-            background: Some(cosmic::iced_core::Background::Color(
-                pressed_theme.on_accent_color().into(),
-            )),
-            ..Default::default()
-        }),
-    }
 }
